+++ nitrite/src/test/java/org/dizitart/no2/DbTestOperations.java
@@ -192,7 +192,7 @@ public class DbTestOperations {
                 ));
         assertEquals(cursor.size(), 2);
 
-        cursor = collection.find(eq("data.1", 4));
+        cursor = collection.find(eq("data.1", (byte) 4));
         assertEquals(cursor.size(), 2);
 
         cursor = collection.find(lt("data.1", 4));
